+++ changelog.md
@@ -1,1 +1,19 @@
-0.0.1 - Nothing
+# Change Log
+
+0.1 - 2019/08/17
+
+- Added readme.md, requires building this out.
+- Added `debugout()` to print out debug log to file in `/debug/` folder with timestamp.
+- Added `webout()` to print web response.text to file in `/out/` folder with timestamp.
+- Added change log for tracking of changes to code.
+- Added road map for tracking of what I intend to do next.
+- Added `scrape()` to use requests command.
+- Added `main()` for running program.
+- Added `inbound()` and `parse()` for handling which sites to connect to.
+
+#### Style
+
+Version - Date version complete - Description
+
+- Change 1
+- Change 2